Fixed go build flag. Moved to a different file.
Added 1 unit test
