+++ Makefile
@@ -43,6 +43,10 @@ release: binaries
 	mkdir -p .release
 	cp .bin/$(NAME)* .release/
 
+.PHONY: test
+test:
+	go test ./... -count=1 -v
+
 .PHONY: integration
 integration:
 	go test --tags=integration ./... -count=1 -v
